Modifications to increase SanchoXEN reliability.
This commit makes functional changes, primarily to the XENduct
infrastructure, to produce a fully reliable XEN based Trusted
Modeling Agent implementation.

The use of an event channel to signal back and forth between the
client and the TMA instance appears to have been a problematic
implementation.  This commit replaces that verification
infrastructure with a boolean value in the shared grant page that
both Sancho and the Quixote reference.

An event channel notification is used to indicate to the receiver
that data is available in the shared memory segment.  The sender
waits on the boolean value that is set by the receiver when the
the grant page contents have been copied into the receive buffer.

The receiver then waits on the boolean signal variable in the
grant page until it is cleared by the sender.

Use of memory barriers is critical to this architecture.  The
code to implement a full memory barrier in the Quixote XENduct
implementation was abstracted from the XEN header files since the
inclusion of the XEN header files containing the barriers was having
dependency issues and the relevant barrier code was a simple one
piece of code.