Change option argument for specifying domain id.
The test-sancho-xen utility was formerly using the -d argument to
specify the domain identifier for the SanchoXen instance that is
to be connected to.

The quixote-xen utility uses the -t argument since the -d
argument is being used by the other utilities to specify that
debug output was to be generated.  This commit converts the test
harness to use the -t argument for specifying the modeling
instance.