Properly respond to clobber target.
The SanchoSGX build infrastructure did not have any commands
implemented for the 'clobber:' target.  The other Sancho
implementations remove the final deliverable, ie. elf image as
the implementation for the clobber command.

This commit makes the clobber target depend on the clean target.
The removal of the signed enclaves was the moved from the clean
target implementation to the clobber target.