Set standard location for STmicro development kits.
The STmicro system development libraries are being packaged into
what will be an installation package that will support the
dependencies needed to compile the MCU based Trusted Modeling
Agents.

The current root location for the SDK's is as follows:

/opt/ESDembedded/share/SDK

Update the build to point to the standard root location for the
development kits.