Remove hash length from identity component digest.
In the previous, IMA based implementation, the SHA256 hash value
representing the COE and Cell identities was computed by the IMA
field hash calculations.  The field hases were pre-pended with
a numberic value representing the size of the hash being used.

The component identities are no longer being generated by the IMA
infrastructure but by the TSEM LSM itself.  Given the difference
in the architecture there is no longer a need to have the length
value included in the hash pre-image.