Consolidate common kernel driver definitions.
The Quixote implementations had grown organically by a great deal
of cut-and-paste, as a result the various definitions for the
kernel driver were scattered across all of the individual Quixote
utility files.

This commit consolidates all of these definitions into a single
file.  This lays the groundwork for making the kernel version be
configurable.

Updates were also conducted to the generic utility documentation
in the top of each source file.  The documentation is in need of
additional improvements.