Convert to TSEM based security point computation.
The TSEM implementation of security modeling uses the following
identity point function:

	point = SHA256(EVENT_TYPE || COE_IDENTITY || CELL_IDENTITY)

Where EVENT_TYPE is a numeric value, starting with 1, that
defines the security event that is being modeled.

The event type number is encoded in the event{} field descriptor.
The previous format of the event{} descript was as follows:

	event{process_name:file_name}

In order to simplify extraction of fields, and to provide the
identity type, the format for the event description is now as
follows:

	event{process=process_name:filename=file_name,type=event_name}

Where process_name is the name, current->comm in Linux, of the
context of execution that is implementing the security event and
file_name is the pathname of the CELL/subject being acted upon.
The event_name is the ASCII text representation of the security
event being executed, ie. file_open, mmap_file etc.

This new method of mapping the security event to a security point
differentiates events that would otherwise have an identical
point representation.  The best example of this is the process of
mapping an executable file that in the new model has two point
representatons, one for the open of the file and one for mapping
of the file.