Update to TSEM based LSM control.
This commit is a large set of changes that updates the Quixote
infrastructure to use the new TSEM based control plane and
namespace infrastructure.

In contrast to the previous Quixote kernel driver the new TSEM
based LSM does not use system calls or the security namespace
that used the generic kernel namespace infrastructure.  These
changes allow the TSEM LSM to be largely independent of the
kernel proper.

With respect to previous system call controls the equivalent
functionality is now implement by writing control messages to the
following pseudo-file:

/sys/kernel/security/tsem/control

With respect to configuration of a TSEM modelspace the following
directives are now supported:

internal

external

enforce

The 'internal' directive creates a security event modeling domain
(modelspace) that is modeled by the in kernel Trusted Modeling
Agent (TMA) implementation.

The 'external' directives creates a security event modeling
domain that is modeled by an external TMA implementation.

The 'enforce' directive configures the modeling domain to return
permission denials if an out of model security event occurs.

The system calls that release a context of execution to run as a
trusted process were replaced by writing the following directives
to the control pseudo-file:

release PID

discipline PID

Where PID is replaced with the process id that is to be acted
upon.

This functionality was all encapsulated in the following new
library object:

TSEMcontrol

In addition to issueing the above modeling domain control
commands the TSEMcontrol object implements the following method:

TSEMcontol->id

This method returns the id number of the security event modeling
domain that the process is running in.  This method is used by
external agents to determine the pseudo-file that will emit the
security events for the domain.

The new TSEM based modelspaces have, as an id number, an unsigned
64-bit integer.  The zero value is reserved to identify the root
modelspace.

The security events for a modeling domain are now surfaced with
the following pseudo-file:

/sys/fs/tsem/update-NNNNNNN

Where NNNNNNN is replaced with the modelspace identity number.
This represents a rename of the root path from
/sys/fs/tsem-events to /sys/fs/tsem.

This update also changes the build order for the SanchoXEN
implementation since the HurdLib library was being built before
the newlib library implementation was built.  This was causing
the HurdLib compilation to fail.

In this commit cycle work was done to verify that the entire
userspace implementation would compile from a pristine source
state.  This uncovered the fact that TE_EVENT to TSEM_EVENT
renames had not been uniformly implemented.  These issues were
fixed for the SanchoSGX and SanchoMCU implementations.

With this commit the following implementations were verified as
working correctly:

Kernel

SanchoUS

SanchoXen

The SanchoSGX and SanchoMCU implementations compile but have not
been verified as direct.  These verifications will be verified
after additional kernel cleanups are implemented that allow the
kernel to be deployed to the Skylake test machine.