Remove redundant event measurement.
The TSEM model object has the call to measure a SecurityEvent
embedded in the TSEM->update method.  As a result the test-TSEM
unit test was failing since that utility was calling the
SecurityEvent->measure method independently of the call to update
the model.

This commit removes that redundant method call to restore
functionality.