Update SecurityEvent object for new format event definition.
The TSEM kernel implementation was updated to base the generation
of a security event coefficient on the following definition:

Sp = SHA256(EVENT_ID || TASK_ID || COE_ID || CELL_ID)

	Where:

		Sp =======> Security point.
		EVENT_ID => uint32_t event number.
		TASK_ID ==> 256 bit task identity.
		COE_ID ===> 256 bit context of execution identity.
		CELL_ID ==> 256 bit cell/inode identity.

Where the TASK_ID is the identity assigned to the task.  This
identity is formed from the same equation above with the
exception that the TASK_ID field is replaced with a 256 bit
(32 byte) buffer filled with null bytes.

The SecurityEvent object was updated to parse the new format of
the event description field that is as follows:

event{proc=process_name, path=pathname, pid=PID, task_id=HEXID}

Where HEXID is a 64 character ASCII hexadecimal encoding of the
256 task identifier.

The SecurityEvent->measure method was updated to feed the updated
coefficient sequence into the compression function that yields
the security event point/identity.

The SecurityEvent->dump method was updated to provide a more
clear output of the components of the security event.

The test-COE, test-cell and test-event unit tests were updated to
embed an updated security event description string.