Implement pseudonym handling for in kernel modeling.
The quixote trust orchestrator did not have support for
implementing pseudonyms.  This commit implements support
for specifying pseudonyms to a kernel based trusted modeling
agent.

Support for writing to the /sys/kernel/security/tsem/pseudonym
pseudo-file was implement in the TSEMcontrol object through the
implementation of the following method:

TSEMcontrol->pseudonym