Output security model parameters for a sealed domain.
Refinements in the TSEM driver now allow a process with CAP_TRUST
to be able to access the pseudo-files that output the security
domain parameters such as the measurement, state, execution
trajectory and forensics trajectory.

There is thus no longer any need to return the null events that
were being previously returned when a sealed model was detected.

This now gives the kernel based TMA the same standing as the
external TMA's with respect to outputing the status of a security
namespace.