Implement security modeling sealing in control file.
The TSEM driver implements the ability to seal a security model
in the /sys/kernel/security/tsem/control pseudo-file, to which
all control plan functionality is being migrated.

The following method was added to the TSEMcontrol object to
enable the implementation of this functionality:

TSEMcontrol->seal

The in kernel TMA implementation of the trust orchestrator was
modified to use this functionality for sealing a security model.