Implement injection of security states through the control file.
The /sys/kernel/security/tsem/map file was dropped from the TSEM
LSM with the functionality replaced by writing the following
directive to the /sys/kernel/security/tsem/control file:

state NNNNNN

Where NNNNNN is replaced with the 256 bit hexadecimal representation
of the security point to be injected into the model.

The following method was added to the TSEMcontrol file to
implement this functionality:

TSEMcontrol->add_state

This allows the add_state function to be dropped from the
quixote.c file.