Add support setting security based with unified control.
Support was added to the TSEM driver for using the unified
control file to set the base point for a security model.  This
commit implements the ability to set the modeling base with the
control file.

Support for doing was implemented in the following object method:

TSEMcontrol->set_base

The TSEM->load method was modified to interpret the following
model definition directive:

base NNNNN

Where NNNNN is the ASCII hexadecimally encoded 256 bit value that
is to be used for the security model base measurement.

The quixote trust orchestrator was modified to interpret the
'base' directive as well.