Add support for socket_create events.
The TSEM driver has support for exporting the parameters of a
socket creation event.  This commit updates the SecurityModel to
handle this event and in the process generalizes the code so that
it is prepared to handle additional non-file based CELL's.

The enumeration that is used to code for the different event
types was moved into a separate header file, tsem_event.h, so
that both the SecurityEvent object and the Cell object have
access to those enumerations.

In order to support communicating the event type to the Cell
object the Cell->parse method was modified to accept the numeric
event type designation as a new argument to the method.

The various tests and utilities were modified to handle this new
support.  A new sample trajectory file that has both file events
and socket_create events was added for testing.