Fix regular expression for socket creation parameters.
The regular expression that was being used for parsing out the
argument to the 'kern' parameter was incorrect.  The parsing
succeeded only because the a non-integer parameter was being
included that resulted in the strtol function terminating
properly.

Fix the regular expression so that it uses a closing brace (})
rather than a comma to terminate the scan of the parameter.