Update security model to use new socket connection format.
The TSEM driver was updated to provide more detailed socket
connection information.  This commit updates the modeling code to
use this new format which provides socket endpoing information
for AF_INET and AF_INET6 sockets.

The format for the sockect connection information is now as
follows:

AF_UNIX:
socket_connect{family=1, addr=DIGEST}

AF_INET:
socket_connect{family=2, port=PORTNUM, addr=ADDR}

AF_INET6:
socket_connect{family=10, port=PORTNUM, flow=FLOW, scope=SCOPE, addr=ADDR}

Where DIGEST is the hexadecimal encoding of a 256 digest of the
pathname of the UNIX socket.

Socket types other than those enumerated above use this same
format.  The difference for these non-codified socket connections
is that the entire data portion of the sockaddr structure is
digested rather than the file pathname.

In the case of the AF_INET and AF_INET6 sockets PORTNUM is the 16
bit port nummer of the endpoing.  ADDR is either the 32 bit IPV4
address or the 128 bit IPV6 address.

FLOW and SCOPE are the integer descriptions of the flow and scope
information from the socket address.

The AF_INET and AF_INET6 values are hard coded in the Cell.c file
secondary to the fact that it simplifies use of the security
modeling code in the embedded modeling environments.

The test-cell utility was updated to provide a more complete
framework for testing Cell parsing.  Arguments are supplied to
the -t command-line switch to specify what type of Cell
definition is to be parsed.