Implement support modeling socket bind events.
This commit implements support for the modeling of a socket bind
event.  Since a socket bind is the receive/accept equivalent of a
socket connection event, the infrastructure for the socket connect
event is largely re-used for computing the mapping of the socket
bind cell.

Various functions in the Cell object were renamed to reflect the
fact that the functions in question have a more general character
than when the implementations were specific to the socket
connection event.

The test-event utility was updated to include a sample IPV6
binding event.  The test is requested by the socket_bind
arguement to the -t command-line switch.