Fix event buffer size.
For the external TMA trust orchestrators there was a mismatch
between the buffer size used for reading the pseudo-file that is
emitting the security event descriptions and the read size of the
pipe that is running in the privileged process.  The former was
using a 512 byte buffer and the latter a 1024 byte buffer.

This caused a regression when support was implemented for more
granular modeling of file_mmap security events.  The emitted
string was being truncated to 512 bytes that was causing a
parsing failure for the event.

Fix this by increasing the size of the buffer used to read
security events to match the size of the buffer used for the pipe
reads.