Implement more advanced modeling of file_mmap events.
The initial implementation of file_mmap modeling was deemed to be
too simplistic, given that it only included the digest of the
file that was being mapped with executable permissions.

The TSEM driver was updated to emit a security event description
of the following form:

mmap_file{type=0, reqprot=N, prot=N, flags=N}

The the type key is used to indicate whether or not this is a
file backed or an anonymous mapping.  A type=1 event is an
anonymous mapping while a type=0 event has a backing file.

The reqprot, prot and flags keys are the arguments passed to the
file_mmap LSM hook.

If the mapping has a file backing a file{} descriptor is also
included in the event description as was the case for the
previous more simplistic implementation.

An additional unit test was added to the test-cell test harness
to test the parsing of a file backed mapping.

The majority of the changes were to the Cell object to implement
proper parsing of the more granular file_mmap implementation.  It
should be noted that a space is needed in front of the prot= tag
in order to prevent it from falsely matching the reqprot= tag.