Rename export event.
In order to rationalize the nomenclature around TSEM modeling the
kernel TSEM driver changed the command directive for the export
of a security event from:

exchange

To:

export

This commit changes the command interpretator table to reflect
this change.  The enumeration type is changed from exchange_event
to export_event.