Implement support for task_kill event modeling.
The TSEM driver was updated to model task_kill events.  This
commit provides support for this modeling in the external
modeling agents.

The parameters for a task kill event are provided by the
following event description field:

task_kill{cross=N, signal=M, task_id=HEXID}

Where N is an integer value that indicates whether or not the
signal is a cross-model call, ie. a signal sent from an alternate
modeling domain.  A value of 0 denotes an intra-domain signal
origination while a value of 1 indates an extra-domain origin.

The M value is the numeric signal number.  HEXID is the
hexadecimal representation of the task identity.