Implement support for socket_accept events.
The Linux TSEM LSM implemented support for modeling socket_accept
events.  This commit provides complementary support for the
external modeling agents to model the events.

The socket accept event is described by the following field
description:

socket_accept{family=N, type=M, port=P, addr=ID}"

Where N is the family number of the socket, M is the type number,
P is the port number and the addr field varies based on the type
of socket.

In the case of an AF_INET socket the address is the hexadecimal
encoding of the 256 bit digest of the socket address.

An AF_INET socket has a number value that is the big-endian
encoded value of the 32-bit internet address.

An AF_INET6 socket has an ASCII hexadecimal encoding of the 128
bit address of the machine representation of the address.