Implement support for flags value in file parameters.
The TSEM driver was updated to include the value of the flags
member of the 'struct file' structure in the file{} event field.
The new parameter is defined by the flags= key/value pair.

This value is important since it defines how the file was opened,
for example, the file denotes whether or not the file was opened
in read/write or read-only mode.