Provide support for ingesting generic events.
The kernel TSEM driver implemented support for generic security
events.  These generic events can be used to rapidly implement
the ingress of a particular event into a model without needing to
implement programmatic support for a more detailed inclusion of
the parameters for the event.

A generic security event is indicated with the following event
type name:

generic_event

The field description is as follows:

generic_event{type=N}

Where N is the enumeration designator for the security event
itself.

The cell value for the event consists of the following function:

cell_mapping = SHA256(type || generic_cell)

Where generic_cell is a 32 byte array of random bytes that was
generated for the kernel driver.  For modeling to be consistent
between the in kernel driver and an external TMA generic cell
identifier must be the same between the external TMA and the in
kernel driver.