Allow explicit naming of process based modeling domains.
Previously the management socket for a process based modeling
domain was named as follows:

/var/lib/Quixote/mgmt/processes/pid-NN

Where NN was replaced with the PID number of the process.

Testing of process based domains proved this to be problematic
since the identity mapping of the AF_UNIX socket that was used to
control the domain was changing for each invocation of the
domain.  This created a situation where deterministic security
models could not be created.

The setup_management() function in each of the trust
orchestrators was modified to use the cartridge name, if not
NULL, as the extension to the pid- string rather than the PID.

So an invocation of a process based modeling domain as follows:

quixote -P -c TEST

Would create a management socket as follows:

/var/lib/Quixote/mgmt/processes/pid-TEST

This allows the establishment of deterministic models to once
again be supported.