Update test-event utility.
The default event string was placed in a CPP define for the
test-event utility in order to take the rather lengthy string out
of the code itself.

The -e command-line argument was introduced to allow a specific
event string to be specified on the command-line to override the
compiled default.