Re-order security event enumerations.
The TSEM driver changed the security enumeration types so that
the TSEM_GENERIC_EVENT was in a slot immediately after
TSEM_BPRM_SET_CREDS.  This stabilizes the security models in the
face of the addition of additional generically modeled events.

This commit modifies the externally modeled agents to use the
same enumeration types.