Fix hang on exit from externally modeled domains.
The introduction of generically modeled events resulted in
the trust orchestrators for generically modeled domains hanging
when the process stack that was being modeled exited.

A very extensive debugging process ended up demonstrating that
the hang was caused by the monitor process (the process that was
reading and forwarding events to the invoked orchestrator)
printing its termination message.  It appears that printing the
message was causing a file_fcntl event to be generated that could
not be handled because the monitor process was no longer
forwarding events to the primary orchestrator process.

This was trivially fixed by not printing the message, which was
arguably of little use.