Move drop of bounding capabilities to forked child process.
Externally modeled domains were hanging and not executing the child
process that would either run a shell or invoke the runc utility
to run a shell in a container environment.

This was caused by a self-referential deadlock by the lead
process attempting to drop its bounding privileges.  The
task_prctl security hook is responsible for approving the
permission to drop the bounding capabilities of a task.  The lead
process is the process that is responsible for receiving security
events from the monitor process and submitting those for approval
to the external modeling agent.

This causes the lead process to hang indefinitely, waiting for
approval for the capability drop that the process itself is
responsible for mediating.

This issue is fixed by moving the bounding capability drop to the
child process that will be invoking either a shell or the runc
binary.  By this time the supervisory infrastructure is in place
to handle approval of the capability drop.