Simplify input loop processing.
Incidental testing of the show-trajectory utility demonstrated
that the event string was continueing to grow in size. This was
secondary to the string not being reset when verbose output was
selected.

An if/then else clause is now used to select the output type.
This allows a single request point for both the SecurityEvent
object and the String object used to hold the input event string
and the re-formatted string..