Implement output of security trajectory and model.
An issue was noted where an exit from an internally modeled
domain resulted in warnings messages being logged about
untrusted security events being intercepted.  In addition, a
subsequent invocation of the utility resulted in an error
secondary to the UNIX domain management socket being left.

This was traced to the fact that using the quixote-console
utility to dump a security model map was not capturing the full
security trajectory.  This was due to the fact that the model was
being captured before the security events involved with the exit
of the utility were processed.

An extensive re-structuring of the quixote utility was conducted
in order to support the export of either a security model map or
the execution trajectory after the modeled process exited.

This required invoking an additional process that 'wraps' the
process that will be handling the monitoring of the modeled
process.  This is an architecture that is now similar to the
externally modeled domains.

Two additional arguments were added to the quixote command-line
syntax to support this functionality.

The -o command-line option is used to specify the name of the
file that the security model or execution trajectory are to be
written.  By default this captures the security model map.

The -t command-line can be added to specify that the execution
trajectory is to be output rather than the securityh model map.