Update userspace orchestrator.
The trust orchestrator that runs the TMA in a userspace process
was updated to add the ability to output a security map or
trajectory in order to match the behavior of the kernel based
trust orchestrator.

The architectural change needed to implement this was different
than the kernel based trust orchestrator since there is no need
to export the model through a pipe back to the launch process.

Trust orchestrators that run in userspace have the access to the
modeling agent in the root modeling domain so the TMA can be
simply queried when the workload ends to output the map or
trajectory.