Output aggregate value from internally modeled domains.
For internally modeled domains the previous behavior was to
output a 'null' aggregate values, ie. 32 null bytes.

With the /sys/fs/tsem/aggregate pseudo-file implemented it is now
possible for an internally modeled domain to have access to the
aggregate value.  Update the general of a security map to use
the aggregate value read from the TSEM filesystem.