Update build for new TPM2 library.
The Quixote_Development package was updated to support the
version 1.6.0 of the IBM TSS library.  In addition to the
standard build of the library a version of the library was
implemented that did not keep state in temporary files.

Modify the build system to support the new library location that
is as follows:

/usr/local/Quixote/lib

And implement minor changes needed to compile and link against
this new library.