Implement utility for creating a TSEM modeling domain.
The following utility was introduced to test the creation of a
TSEM modeling domain:

test-domain-creation

This is a very simple utility that forks a process that creates a
modeling domain with the parent process waiting for the child
that created the domain to terminate.  The child issues a pause()
system call and can be terminated by sending any signal to it.

This utility provides a controlled mechanism for debugging domain
creation that is more tractable and capable of yielding better
forensics than running a trust orchestrator.