Implement support for building custom libcap library.
One of the issues with supporting multiple kernel verions with
Quixote is the fact that each supported kernel needs a different
libcap library in order to handle the CAP_TRUST capability.

This commit introduces the Support directory in the top of the
source tree that will be used to support this library and any
additional libraries that need to be downloaded and/or built.

The Makefile in the Support directory uses a framework that was
developed for the SanchoXen TMA for downloading and building
support libraries.

The Support directory was configured to download the most current
release of the libcap library.  The patch commited with this
update configures that library to build against a TSEM enabled
6.1 kernel release.

This commit also introduces the beginning of which will be a
global build configuration withe the Config.mak file in the
top-level directory.  The Makefile in the Support directory was
modified to either respect the parameters that are passed down or
to source the configuration file directory.

The following make variable is exported to signal that a global
build is being conducted:

GLOBAL_BUILD = true

This configuration file initially supports the ability to select
a compiler and a kernel version to be supported.