Convert to using event names rather than numeric types.
The TSEM driver converted to the use of the security event name
rather than the numeric enumeration type that was linked to the
name.  This was after a consideration of how to maintain a stable
ABI from the TSEM driver to userspace.

This commit updates the external modeling code to follow this
strategy.  The most principal change is to conver the Cell object
to extract the event name in the form of a String object in the
state information for the Cell.

For generically modeled events the type= parameters of the
generic_event field also uses the name of the event rather than
the enumeration type.  As a result the show-trajectory utility is
no longer needed since a trajectory entry can be interpreted
directly.  This allows the format_generic methods to be removed
from the SecurityEvent and Cell object.