Fix TPM2 utility build against new IBMTSS libraries.
With the move of the IBMTSS libraries to the /usr/local/Quixote
directory heirarchy the compilation of the tpm2-cmd utility was
failing.

Update the Makefile to reflect a correct include directory to be
searched to fix the include problem.

It appears that the updated IBMTSS API no longer needs to include
the tssproperties.h file so it was removed from the include list.