Extend global build configuration architecture.
This commit updates the build infrastructure to provide a
framework for having a site or build configuration file that is
used to override any parameters that are in the Config.mk file in
the top build directory.

In order for build to be implemented from sub-directories the
following command must be run in the root of the source tree:

make config

This will create a sample Build.mk file which can be edited to
implement the site overrides for the build.

Makefiles in subordinate directories execute the following in
order to source the build configuration:

ifndef BUILD_CONFIG
include ${shell cd PATH_TO_ROOT; pwd}/Build.mk
endif

Where PATH_TO_ROOT is replaced with a relative path declaration
to the root of the build tree.  For example a directory
immediately below the top level would implement the following
command:

ifndef BUILD_CONFIG
include ${shell cd ..; pwd}/Build.mk
endif

It is anticipated that there will be other significant changes to
the build process for the next release, primarily consolidation
of variables and rules that are now replicated through multiple
sub-directories.