Eliminate compiler dependency for SanchoXen TMA.
The build system for the SanchoXen TMA had been hard coded to use
the directory name for the compiler specific library directory to
set the location for the libgcc.a file.

Fix this by using the -print-libgcc-file-name to output the
location of the library file that contains the GCC intrinsics.
The directory for that library file is the directory that holds
the directory that contains the platform include headers.

Use the directory location for the libgcc file to define the
directory for the platform include files.

This strategy will make the build independent of the version of
the compiler being used.