Update untrusted event log handling.
The TSEM driver updated how messages about untrusted tasks
executing security event handlers were being exported.

The keyword for the export changed from:

TSEM_event

To the following:

log

The COE{} field was being used to define the name of the process
that was requesting the event.  This was effectively a class with
the export event that was using the COE{} in the form that is
used for the output of trajectory and forensics values.

The name of this field was changed from the following:

COE

To:

process

Update the trust orchestrators to use this new syntax.