Generalize naming of libcap patches.
Given that userspace will support multiple TSEM kernel
implementations the strategy of using the following format for
the libcap patches was flawed:

libcap-MAJOR.MINOR_KERNEL.patch

Where MAJOR/MINOR were the libcap major and minor releases and
KERNEL was replace with the kernel version, ie. 5.4.

The change to the libcap patches is secondary to what bit number
the kernel will be using for the CAP_TRUST capability.  For
example, the 5.15 LTS and 6.x LTS kernel versions both have the
CAP_TRUST bit as 41.

The format for naming the patches was thus changed to the
following:

libcap-MAJOR.MINOR_cap_BITNUM.patch

Where MAJOR/MINOR retain the same definitions as noted above and
BITNUM is the bit number assigned to the CAP_TRUST CAPABILITY.