Update SanchoXen based trust orchestrator.
The quixote-xen utility that runs a Trusted Modeling Agent in a
Xen hypervisor stub domain, was updated to support the new
workload monitoring architecture that conducts the monitoring in
a separate function.

Support was also added to output either a security model or
security execution trajectory at the end of the execution of the
modeled execution domain.

The utility for testing the SanchoXen implementation was modified
to use the -s command-line argument to specify the domain number
that will be running the TMA instance.