Fix compile warnings in userspace trust orchestrator.
The upgraded GCC 9.5 compiler is warning for code anomalies that
the previous compiler was not.

One of the issues that had crept into the codebase for the trust
orchestrators was comparing the TSEMcontrol method returns to a
return value rather than testing for a boolean return value.
Remove the integer comparison and simply test the return value as
a boolean value.

The child_monitor function was place the return code from the
pipe read into the boolean return variable, retn.  Conver this
code to use the integer rc variable and set the function return
value based on whether or not an error was encountered.