Re-order inclusion of local build directives.
Testing was indicating that the setting of the C compiler was
being done by the default CC definition supplied by the compiler.
This caused a re-evaluation of how to handle the relationship
between the build configuration and local modifications to it.

The inclusion of the local configuration file was moved to after
the defaults were set.  This is more logically intuitive given
that the local build configurations should override the global
defaults.