Fix error return handling.
Testing indicated that the SanchoXen based trust orchestrator was
dumping an execution stack trace, secondary to an error being
generated when a cartridge workload terminates.

This was traced to the fact that the return code for the
child_monitor() function was not having its return value properly
set.

A review of the code prompted a minor change in how the return
code is generated.  An else clause was added to the conditional
that checks if a security model or trajectory is to be output.
The else clause sets the return value to true, given that the
condition is executed after the child_monitor() function has
returned without error.

Update the SanchoUS userspace trust orchestrator to use the same
strategy.