Add TSEM LSM kernel documentation.
Add the file that is included in the Linux kernel source tree to
the Documentation directory.  The file is in the 'Sphinx' .rst
format but is readable as standard ASCII text.

Updates to the documentation will be done here as the kernel
source is transitioned to the LSM branch that will be used to
upstream the TSEM driver.